chore(primitives): move getTestId util
