+++ packages/react/src/primitives/utils/getTestId.ts
@@ -1,0 +1,5 @@
+export const getTestId = (
+  testId?: string,
+  component?: string
+): string | undefined =>
+  testId && component ? `${testId}-${component}` : undefined;
